fix: Re-add favicon assets since they are used on the forum
The forum uses the asset from GitHub as favicon. This asset
was removed from the repository since it is the stack now
that serves the favicon (see [3c9051a](https://github.com/cozy/cozy-home/commit/3c9051aee3de27df7522b9604d788321ffb07b20)).

For the forum to have an image, as a temporary action, this
commits puts back the icon in the repository until we find
a better place to store static assets for the forum.