feat: Remove unsupported locales from the bundled JS
We had in the bundle a log of JSON and JS for locales that we don't
support. It takes a lot of spaces (several hundreds of kb) and it is not
used as it's not possible to select one of those locales currently. So,
let's remove them to make the home loads faster.