feat: Handle cozy-harvest-lib@0.33.0 ✨
This version of cozy-harvest-lib introduces error
management and get rid of the Success view after
account configuration. So Cozy-Home manage to
show this view.

It implies that Home does not longer pass an onDone
hook but a onSuccess instead.

onSuccess correspond to a fully done konnector job,
where onLoginSucccess correspond to a login success.