feat: Fetch trigger after job update 📝
Now that Harvest is launching the triggers, Cozy-Home
must stay up-to-date. Fetching trigger after a job
update allows us to have the error associated to
the trigger and pass it to Harvest's TriggerManager.

Without these change an error on account creation
is not displayed immediately when editing account
right after.