fix: Take banks url from store inside context
A mistake from my end ended up costing many hours of my life.
The second parameter of mapDispatchToProps is *not* the store,
but the ownProps. The simple way to access CollectStore is
through this.context. The best way would be to access it with
a ContextProvider. This should be done when home does not use
pReact