refactor: Add .styl extension on stylus imports
Tests were replacing all imports with `styles` in the import path by
`identity-obj-proxy`. But this rule was too generic, and when a
dependency comes with an import path in which we can find `styles`, we
were also replacing it. But it's not because there is `styles` in the
path that it's not a JS module. Adding the `.styl` extensions allows us
to remove this too generic rule and only target the import paths that
end with `.styl`