refactor(KonnectorInstall): Remove useless onCancel prop
This prop was used in the context of a previous vault unlock form design
which had a "close" button. In that case we are to propagate the action
to parent components using the `onCancel` prop. But this button doesn't
exist anymore and the vault unlock form modal is not closable. I added a
test that ensures we pass the `vaultClosable` prop to the TriggerManager.