fix: Prevent flow crash when adding a banking account from intent
Since banking account communicate directly with BI, when the
LOGIN_SUCCESS event is triggered, the trigger that is dispatched does
not contain an id (it is fake trigger, the trigger has not yet been
created). This would make the enqueueConnection action handling
crash with "Missing trigger id" problem.

Ignoring the action fixes the problem.