fix: Fallback to slug name in URL
When there is no Konnector prop, the application will crash.
This happens because no slug is passed to Konnector component.
In order to fix this, we can read the URL and get the slug from there.
As URL are predictible, we can safely assume the slug will always be
the third string in the hash pathname.