fix: Intent redirect
If we came from the store, by clicking on "open" on a konnector,
we're redirected to the Home application with the /redirect path

Inside this redirect path, we render the IntentRedirect component.
This component reads the redux store to fetch the current installed
konnectors.

But since we were setting isFetching to false by default, we had, at the
first render, isReady to true. So we redirected to /redirected without
having the informations about the konnectors yet.

The fix is pretty simple. We need to check if at the first render, one
query is in loading state.

Also, I refactored to use `some` instead of `find` to only deal with
boolean.