fix: Perf optimization
We have a repaint when we add the `app-opened` class. We should not have
one, but I didn't find yet why we have one (maybe because of the
backdrop-filter but not sure).

So by adding this will-change transform, we say to the browser that
it can create a new layer for those 2 nodes. Like that, it'll only
repaint them.

See https://web.dev/simplify-paint-complexity-and-reduce-paint-areas/#promote-elements-that-move-or-fade
for more informations