fix: Use `RemoveScroll` to prevent OS pull-to-refresh on Flagship app
In #1730 we attempted to fix the background's position so it would not
show a white area on top of the app when trying to pull-to-refresh from
our Flagship app

This had 2 side effects

The first one is that pull-to-refresh was not possible anymore on iOS
Safari and also the address bar would not hide anymore when scrolling

To prevent this we chose to apply the custom CSS only when the app is
hosted by our Flagship app

The second side effect is that trying to trigger the pull-to-refresh
gesture would freeze the UI for a few seconds. This is a common issue
with `position: fixed` divs

To prevent this, we wrapped the `main` div with `<RemoveScroll />`
plugin that would completely lock the scroll of the HTML body and only
allow the `main` div to scroll. However the `position: fixed` is still
needed si the `main` div would have the correct height and enable
scroll