feat: Add cliskTimeout service
This service stops the given job with a 'context deadline exceeded' error. It may be called when
the execution of the clisk konnector is too long or when the flagship app has crashed.
The flagship app will also timeout the job if it restarts but this way, the user will not see
the job as running even outside the flagship app

This service is supposed to be called directly or via a @in or @at trigger

Example to call this service :

```javascript
await client.collection('io.cozy.jobs').create(
   'service',
   {
     name: 'cliskTimeout',
     slug: 'home',
     fields: {
      cliskJobId: '<the job id>'
     }
   }
 )
```