fix: prevent double render in dev mode
The app was always rendered twice in a row with a 30ms interval
This was because renderApp() was called twice with bad
HMR handling.
HMR update has been disabled for now because it doesn't work correctly.
The whole page will reload just as before when a file is updated.
This will  be fixed in another commit