feat: Adapt background to current theme
We want our partners to be able to have more control on the displayed
background

In combination with the new `--home-theme` variable (see previous
commit) a partner can customize the Cozy's background by targeting
`home-default-partner-background` css class on the custom CSS from
dynamic assets. However this can only be done when choosing the
`normal` theme

To ease background customization, the background's blend mode is
deactivated when `--home-theme` is set to `normal`. But the partner can
add it back in the `home-default-partner-background` css class

Note that the `background` props should be forced using `!important`
CSS flag or the Cozy's JPG background will be prioritized instead

Note that if the user sets their own custom background, then the
`home-default-partner-background` css class will be ignored as for the
`--home-theme` value (see previous commit). The background's blend mode
will also be reapplied