fix: Use `trim()` on `usePreferedTheme`
In some environments `getComputedStyle().getPropertyValue()` would
return a result with a leading space

In this scenario we retrieve ` normal` instead of `normal`

So we want to `trim()` the result

Note that `getComputedStyle().getPropertyValue()` always return a
string even if the property does not exist. In that cas the string is
empty. So we don't need to check for nullity