fix: Normal theme for Logout Modal
Cozy Home can have normal or inverted theme now,
so we need to specify the theme we want when we
open a modal.

In this case, let's force the Normal theme.

We had to force the position absolute to the theme
because CozyTheme force a div block and it breaks
the flow of the buttons.