feat: Reload data on focus
Since we heavily rely on relatime to update the redux store, we
can miss a message when the app is in background or else. It can
result to open Harvest with outdated information and it will then
display an error.

Here, we refresh triggers, accounts and konnectors when the focus
is there.

This code will be removed once the migration to cozy-client is
done.