fix: render null if konnector not in registry
If there is no registry data it is better to render nothing
because we won't have any link to the Konnector,
despite having its slug.
This is because failing to get the data most likely means
that the Service is not available
on the current registry (404 Not Found).
It is not a problem if the Service is not available on the registry,
but it is a problem if we try to link to it.
Here it would display an AppLinker with a grey cube,
an unformatted slug as a name and no working link.