fix: Use Cozy-Client store for KonnectorTile
We had a few issues with the data persistence within the flagship app.
This is mostly due to the fact that we have a lot of way to manage data
for the Home application.

Here, I remove the `fetchTriggers` to use query(Q('io.cozy.triggers'))
everywhere within the app.

I also changed the konnectorTile to use data from the cozy-client store
instead of the one from the Home.