chore: Disable experimental fetch
I had one error when upgrading to node 18 :
  ● polyfillFetch › when fetch and related objects does not exist in global scope › should use polyfills

    TypeError: Cannot assign to read only property 'Headers' of object '[object global]'

It seems related to the fetch() method enable by defaut by node18
which relies on `undici`. And `undici` seems to add Headers, so we
can't write them.

By disabling this fetch api, the issue is resolved. I don't think
this should be a long run fix, but it'll make the migration easier.