fix: Request the shortcutsDirectory at the top level
Like that, we can wait this request to be done in order to hide the
splashscreen.

Before this fix, we were sending too soon the message, and on an
application with a lot of shortcut, the result for the user was not
good.

Here we wait for all the info to be fetched before setting the isReady