feat: Add VIEW INTENT
The idea is to be able to display an account from every where.
This intent is not directly used in production, but it's useful to test
the end (aka the display of harvest) of the intent
create io.cozy.accounts.

I used it to debug the "open in Drive" feature, since the link was not
working.