feat: Add a context to store the custom Wallpaper
We already reduce the number of request to wallpaper, but here we go
further and make only one call (see https://github.com/cozy/cozy-home/pull/2020
for the first iteration).

Also, we store 2 things :
- within the localStorage : the fact that we have a custom wallpaper or
not, in order to have a quick path to know what to do
- within the localforage, the binary of the wallpaper. I decided to store
it within localforage, because localstorage has a limite to 5 or 10MB
depending on the browser. And since a wallpaper can be huge... like that
"no limit".

Maybe the next iteration will be to call "hidesplashscreen" only if the
wallpaper is fully loaded.