fix: Home Style
Since 5c850137f7a26fa1b35576408cf93f5cfbc4f01e I store the
default wallpaper into the wallpaperlink var of the hook.

But I didn't expect that an other hook was relying on this
value to determine if it was an inverted theme or not.

So I fix the issue by checking if the current wallpaper
is different of the default one.