fix: Call isReady
In a previous attempt fda23d869ecd63ef6805ec171d911c78ee9d3928 I managed
to try to fix the call to hideSplashScreen.

I fixed the call when the user had a magicFolder but without custom
shortcuts. But missed the case where the user doesn't have a magic
folder.

This is a TMP fix. I need to think about a better fix and mainly a way
to add a few UT to that method.