fix: Avoid to make the background image too dark in dark mode
We use a combination of backdrop-filter and mix-blend-mode to darken the
background image. However, this take into account the background color
of the body which is white in light mode and a strong gray in dark mode.

So here we override the background color in .App to white.
It has no effect on the container using .App, but it avoid
mix-blend-mode to be linked to the background color of the body.