fix(groupModeAction): Make the full line clickable
Before, only way to de-group was to click on the  switch. Now it's the line in the options menu that does the action (the switch is contained inside).

I've taken advantage of this to make the user return to the root when he chooses to ungroup a folder so that he doesn't stay on a route that doesn't display anything. This could lead to strange behaviour when they chose to group it again