feat: Improve triggers query performances
We used to query all triggers, in order to determine which konnectors
accounts are connected, to adapt display on the homepage.

However, this query can be very time-consuming: it takes ~10s on an
instance with 700+ triggers. This is due to the stack computing job
information from each trigger.
As we do not need such information here, we force the generic `/data`
route, rather than the specialized `/jobs/triggers` route.