feat: Simplify and speed up Home initialization
We get rid of the  complicated state management to determine whether or not the
Home is initialized, to rely on a dedicated hook in charge of fetching all the required
data.
Note the Home is relying on this mechanism to later fetch data directly from the store.
In the future, we might have the same logic, but relying on the dataproxy directly