fix: Get triggers current_state efficiently
We used to query triggers with their `current_state` at Home startup. We
removed it for performances purposes, but now it is only made when the
focus event is fired. This means we can miss this state when entering a
konnector group, which is required to correctly display the icons.
So, we query the triggers when required, and mutualize it with the focus
queries to benefit from the redux store.