fix: Prevent infinite loop at startup
At startup we check some "app highlights" status, typicallky for backup
or geolocation tracking.
In the tracking case, we need to query the database in order to check if
there is some geolocation data.
This query might fail in offline if the local database is not ready.

And this failure was causing an infinite loop, because a retry was made
for each failed query.