fix: Avoid rendering at focus
At each new focus, a full render was triggered, making a visual refresh
for the user.
To fix this, we do not use query options in the ReloadFocus component,
because of: https://github.com/cozy/cozy-client/issues/931
Especially for the apps query, the fetchPolicy is always applied here,
because it is called elsewhere, making the query's lastUpdate very close
in time. And because of this, the home is rendered after each focus, as
the same query gives different result (a complete response, vs an early
empty return in this case)