feat: Do not translate anymore home magic folder
We were looking for a translated name for home magic folder, but
translations were always the same except for deutsch, spanish ("V"
???) and russian (a language added in the last days).

So let's remove these translations that only complicates things for nothing. If later we want translated name for home settings folder, we will use a real magic folder.