refactor: Improve AddTile organization
AddTile can do two things depending of multiple factors:
- open shortcut modal
- open store app

In dcf5018c95d4caf3a14ae14573bf2921bb3492be, we moved down the
isStoreAvailable check that was previsouly conditioning the
visibility of AddTile inside AppTile. Without this, if the store was
not installed, even if we wanted "open shortcut modal" behavior, it
was not possible because the AppTile component was never mounted.

Here I separate AppTile in two subcomponents and properly check which
one should be displayed in the correct order in the AppTile which
become a wrapper.